requirements.txt
mysql-connector-python>=8.3.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
